+++ src/lib.rs
@@ -13,3 +13,4 @@ pub mod rfc4791;
 pub mod rfc4918;
 pub mod rfc5397;
 pub mod rfc6352;
+pub mod rfc6578;